+++ CHANGELOG.md
@@ -3,7 +3,14 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [10.7.0] 2025-10-07
+
+### Fixed
+
+- ML_datasetmaster: internal training and test datasets based on `_PytorchDataset` will correctly inherit the feature names and target names from the classes using them.
+- ML_trainer: `MLTrainer.explain()` correctly handles Datasets with feature names and target names as needed.
+- ML_scaler: `PytorchScaler.save()` properly handles filepaths.
 
 ## [10.6.0] 2025-10-07